fix: Move `cpuid` into module under `vmm`
