--- src/cpuid/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "cpuid"
-version = "0.1.0"
-authors = ["Amazon Firecracker team <firecracker-devel@amazon.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-kvm-bindings = { version = "0.6.0", features = ["fam-wrappers"] }
-kvm-ioctls = "0.12.0"
-thiserror = "1.0.32"
-
-arch = { path = "../arch" }
-arch_gen = { path = "../arch_gen" }
-utils = { path = "../utils"}
