+++ src/vmm/Cargo.toml
@@ -32,9 +32,7 @@ snapshot = { path = "../snapshot"}
 utils = { path = "../utils" }
 virtio_gen = { path = "../virtio_gen" }
 vm-memory = { path = "../vm-memory" }
-
-[target.'cfg(target_arch = "x86_64")'.dependencies]
-cpuid = { path = "../cpuid" }
+arch_gen = { path = "../arch_gen" }
 
 [dev-dependencies]
 criterion = "0.4.0"
