+++ src/vmm/src/cpuid/cpu_leaf.rs
@@ -6,7 +6,7 @@ pub mod leaf_0x1 {
     pub const LEAF_NUM: u32 = 0x1;
 
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         pub const EXTENDED_FAMILY_ID_BITRANGE: BitRange = bit_range!(27, 20);
         pub const EXTENDED_PROCESSOR_MODEL_BITRANGE: BitRange = bit_range!(19, 16);
@@ -17,7 +17,7 @@ pub mod leaf_0x1 {
     }
 
     pub mod ebx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The bit-range containing the (fixed) default APIC ID.
         pub const APICID_BITRANGE: BitRange = bit_range!(31, 24);
@@ -77,7 +77,7 @@ pub mod leaf_0x1 {
 
 pub mod leaf_cache_parameters {
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         pub const CACHE_LEVEL_BITRANGE: BitRange = bit_range!(7, 5);
         pub const MAX_CPUS_PER_CORE_BITRANGE: BitRange = bit_range!(25, 14);
@@ -89,9 +89,9 @@ pub mod leaf_0x4 {
     pub const LEAF_NUM: u32 = 0x4;
 
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
         // inherit eax from leaf_cache_parameters
-        pub use crate::cpu_leaf::leaf_cache_parameters::eax::*;
+        pub use crate::cpuid::cpu_leaf::leaf_cache_parameters::eax::*;
 
         pub const MAX_CORES_PER_PACKAGE_BITRANGE: BitRange = bit_range!(31, 26);
     }
@@ -236,7 +236,7 @@ pub mod leaf_0xb {
     pub const LEVEL_TYPE_CORE: u32 = 2;
 
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The bit-range containing the number of bits to shift right the APIC ID in order to get
         // the next level APIC ID
@@ -244,7 +244,7 @@ pub mod leaf_0xb {
     }
 
     pub mod ebx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The bit-range containing the number of factory-configured logical processors
         // at the current cache level
@@ -252,7 +252,7 @@ pub mod leaf_0xb {
     }
 
     pub mod ecx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         pub const LEVEL_TYPE_BITRANGE: BitRange = bit_range!(15, 8);
         pub const LEVEL_NUMBER_BITRANGE: BitRange = bit_range!(7, 0);
@@ -265,7 +265,7 @@ pub mod leaf_0xd {
 
     pub mod index0 {
         pub mod eax {
-            use crate::bit_helper::BitRange;
+            use crate::cpuid::bit_helper::BitRange;
 
             pub const MPX_STATE_BITRANGE: BitRange = bit_range!(4, 3);
             pub const AVX512_STATE_BITRANGE: BitRange = bit_range!(7, 5);
@@ -286,7 +286,7 @@ pub mod leaf_0x80000000 {
     pub const LEAF_NUM: u32 = 0x8000_0000;
 
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         pub const LARGEST_EXTENDED_FN_BITRANGE: BitRange = bit_range!(31, 0);
     }
@@ -330,7 +330,7 @@ pub mod leaf_0x80000008 {
         pub const IBRS_PROVIDES_SAME_MODE_PROTECTION_BITINDEX: u32 = 19;
     }
     pub mod ecx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The number of bits in the initial ApicId value that indicate thread ID within a package
         // Possible values:
@@ -348,7 +348,7 @@ pub mod leaf_0x8000001d {
     pub const LEAF_NUM: u32 = 0x8000_001d;
 
     // inherit eax from leaf_cache_parameters
-    pub use crate::cpu_leaf::leaf_cache_parameters::eax;
+    pub use crate::cpuid::cpu_leaf::leaf_cache_parameters::eax;
 }
 
 // Extended APIC ID Leaf
@@ -356,13 +356,13 @@ pub mod leaf_0x8000001e {
     pub const LEAF_NUM: u32 = 0x8000_001e;
 
     pub mod eax {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         pub const EXTENDED_APIC_ID_BITRANGE: BitRange = bit_range!(31, 0);
     }
 
     pub mod ebx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The number of threads per core - 1
         pub const THREADS_PER_CORE_BITRANGE: BitRange = bit_range!(15, 8);
@@ -370,7 +370,7 @@ pub mod leaf_0x8000001e {
     }
 
     pub mod ecx {
-        use crate::bit_helper::BitRange;
+        use crate::cpuid::bit_helper::BitRange;
 
         // The number of nodes per processor. Possible values:
         // 0 -> 1 node per processor
