+++ src/vmm/src/cpuid/template/amd/t2a.rs
@@ -3,16 +3,16 @@
 
 use kvm_bindings::{kvm_cpuid_entry2, CpuId};
 
-use crate::bit_helper::BitHelper;
-use crate::cpu_leaf::*;
-use crate::template::amd::validate_vendor_id;
-use crate::transformer::*;
+use crate::cpuid::bit_helper::BitHelper;
+use crate::cpuid::cpu_leaf::*;
+use crate::cpuid::template::amd::validate_vendor_id;
+use crate::cpuid::transformer::*;
 
 fn update_extended_feature_info_entry(
     entry: &mut kvm_cpuid_entry2,
     _vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0x80000001::*;
+    use crate::cpuid::cpu_leaf::leaf_0x80000001::*;
 
     entry
         .ecx
@@ -36,7 +36,7 @@ fn update_extended_feature_extensions_entry(
     entry: &mut kvm_cpuid_entry2,
     _vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0x80000008::*;
+    use crate::cpuid::cpu_leaf::leaf_0x80000008::*;
 
     entry
         .ebx
@@ -55,9 +55,9 @@ struct T2ACpuidTransformer;
 impl CpuidTransformer for T2ACpuidTransformer {
     fn entry_transformer_fn(&self, entry: &mut kvm_cpuid_entry2) -> Option<EntryTransformerFn> {
         match entry.function {
-            leaf_0x1::LEAF_NUM => Some(crate::t2::update_feature_info_entry),
-            leaf_0x7::LEAF_NUM => Some(crate::t2::update_structured_extended_entry),
-            leaf_0xd::LEAF_NUM => Some(crate::t2::update_xsave_features_entry),
+            leaf_0x1::LEAF_NUM => Some(crate::cpuid::t2::update_feature_info_entry),
+            leaf_0x7::LEAF_NUM => Some(crate::cpuid::t2::update_structured_extended_entry),
+            leaf_0xd::LEAF_NUM => Some(crate::cpuid::t2::update_xsave_features_entry),
             leaf_0x80000001::LEAF_NUM => Some(update_extended_feature_info_entry),
             leaf_0x80000008::LEAF_NUM => Some(update_extended_feature_extensions_entry),
             _ => None,
