+++ src/vmm/src/cpuid/template/intel/t2cl.rs
@@ -4,16 +4,16 @@
 use arch::x86_64::msr::{ArchCapaMSRFlags, MSR_IA32_ARCH_CAPABILITIES};
 use kvm_bindings::{kvm_cpuid_entry2, kvm_msr_entry, CpuId};
 
-use crate::bit_helper::BitHelper;
-use crate::cpu_leaf::*;
-use crate::template::intel::validate_vendor_id;
-use crate::transformer::*;
+use crate::cpuid::bit_helper::BitHelper;
+use crate::cpuid::cpu_leaf::*;
+use crate::cpuid::template::intel::validate_vendor_id;
+use crate::cpuid::transformer::*;
 
 fn update_extended_feature_info_entry(
     entry: &mut kvm_cpuid_entry2,
     _vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0x80000001::*;
+    use crate::cpuid::cpu_leaf::leaf_0x80000001::*;
 
     entry
         .ecx
@@ -39,11 +39,13 @@ struct T2CLCpuidTransformer;
 impl CpuidTransformer for T2CLCpuidTransformer {
     fn entry_transformer_fn(&self, entry: &mut kvm_cpuid_entry2) -> Option<EntryTransformerFn> {
         match entry.function {
-            leaf_0x1::LEAF_NUM => Some(crate::t2::update_feature_info_entry),
-            leaf_0x7::LEAF_NUM => Some(crate::t2::update_structured_extended_entry),
-            leaf_0xd::LEAF_NUM => Some(crate::t2::update_xsave_features_entry),
+            leaf_0x1::LEAF_NUM => Some(crate::cpuid::t2::update_feature_info_entry),
+            leaf_0x7::LEAF_NUM => Some(crate::cpuid::t2::update_structured_extended_entry),
+            leaf_0xd::LEAF_NUM => Some(crate::cpuid::t2::update_xsave_features_entry),
             leaf_0x80000001::LEAF_NUM => Some(update_extended_feature_info_entry),
-            leaf_0x80000008::LEAF_NUM => Some(crate::t2::update_extended_feature_extensions_entry),
+            leaf_0x80000008::LEAF_NUM => {
+                Some(crate::cpuid::t2::update_extended_feature_extensions_entry)
+            }
             _ => None,
         }
     }
