+++ src/vmm/src/cpuid/transformer/intel.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use super::*;
-use crate::bit_helper::BitHelper;
-use crate::cpu_leaf::*;
+use crate::cpuid::bit_helper::BitHelper;
+use crate::cpuid::cpu_leaf::*;
 
 // The APIC ID shift in leaf 0xBh specifies the number of bits to shit the x2APIC ID to get a
 // unique topology of the next level. This allows 128 logical processors/package.
@@ -13,7 +13,7 @@ fn update_deterministic_cache_entry(
     entry: &mut kvm_cpuid_entry2,
     vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0x4::*;
+    use crate::cpuid::cpu_leaf::leaf_0x4::*;
 
     common::update_cache_parameters_entry(entry, vm_spec)?;
 
@@ -30,7 +30,7 @@ fn update_power_management_entry(
     entry: &mut kvm_cpuid_entry2,
     _vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0x6::*;
+    use crate::cpuid::cpu_leaf::leaf_0x6::*;
 
     entry.eax.write_bit(eax::TURBO_BOOST_BITINDEX, false);
     // Clear X86 EPB feature. No frequency selection in the hypervisor.
@@ -54,7 +54,7 @@ fn update_extended_topology_entry(
     entry: &mut kvm_cpuid_entry2,
     vm_spec: &VmSpec,
 ) -> Result<(), Error> {
-    use crate::cpu_leaf::leaf_0xb::*;
+    use crate::cpuid::cpu_leaf::leaf_0xb::*;
 
     // reset eax, ebx, ecx
     entry.eax = 0_u32;
@@ -160,8 +160,8 @@ mod tests {
     use kvm_bindings::kvm_cpuid_entry2;
 
     use super::*;
-    use crate::cpu_leaf::leaf_0xb::{LEVEL_TYPE_CORE, LEVEL_TYPE_THREAD};
-    use crate::transformer::VmSpec;
+    use crate::cpuid::cpu_leaf::leaf_0xb::{LEVEL_TYPE_CORE, LEVEL_TYPE_THREAD};
+    use crate::cpuid::transformer::VmSpec;
 
     #[test]
     fn test_process_cpuid() {
@@ -207,7 +207,7 @@ mod tests {
         cache_level: u32,
         expected_max_cores_per_package: u32,
     ) {
-        use crate::cpu_leaf::leaf_0x4::*;
+        use crate::cpuid::cpu_leaf::leaf_0x4::*;
 
         let vm_spec = VmSpec::new(0, cpu_count, smt).expect("Error creating vm_spec");
         let entry = &mut kvm_cpuid_entry2 {
@@ -239,7 +239,7 @@ mod tests {
         expected_num_logical_processors: u32,
         expected_level_type: u32,
     ) {
-        use crate::cpu_leaf::leaf_0xb::*;
+        use crate::cpuid::cpu_leaf::leaf_0xb::*;
 
         let vm_spec = VmSpec::new(0, cpu_count, smt).expect("Error creating vm_spec");
         let entry = &mut kvm_cpuid_entry2 {
